fix loop echo
