+++ src/commands/duty/next430.js
@@ -1,29 +1,17 @@
-const db = require('./db')
-const personList = require('./personList')
 const to430 = require('./to430')
-const isHoliday = require('./is-holiday')
+const echo = require('./echo')
 const getNow = require('../../utils/now')
-
-const personSize = Object.keys(personList).length
+const isHoliday = require('./is-holiday')
 
 function next430() {
   const to430ms = (to430() - getNow())
   const hours24ms = 24 * 60 * 60 * 1000
 
   const loopin24H = ()=> {
-    if (isHoliday()) return
+//    if (isHoliday()) return setTimeout(loopin24H, hours24ms)
 
-    db.findOne({}, (er, doc)=> {
-      const dutyID = doc.dutyID
-      const nextDutyID = dutyID + 1
-      const who = personList[dutyID]
-
-      // Cepave all: -23470446
-      ai.sendMessage(-23470446, `今天的值日生是: ${who.name} (${who.en})`)
-
-      db.update(doc, {dutyID: nextDutyID > personSize ? 1 : nextDutyID})
-    })
-
+    // Cepave all chadId: -23470446
+    echo(-23470446, true)
     setTimeout(loopin24H, hours24ms)
   }
