+++ src/commands/duty/person-list.js
